pyproject.toml
[project]
name = "pwn-college"
version = "0.1.0"
description = "My attempts to the pwn.college challenges]"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dojjail>=0.1.7",
    # pwntools 4.14.1 requires unicorn>=2.0.1, !=2.1.3, !=2.1.4
    # which having trouble installing on Windows, so we force a 'broken' version
    # https://github.com/unicorn-engine/unicorn/issues/2245
    # https://github.com/unicorn-engine/unicorn/issues/2134
    "flask>=3.1.3",
    "pwntools==4.15.0",
    "pycryptodome>=3.23.0",
    "rich>=15.0.0",
    "scapy>=2.7.0",
    "unicorn==2.1.1",
    "watchdog>=6.0.0",
]

[tool.uv]
# this project is not for packaging.
package = false

[tool.pyright]
# assign sys.platform to system challenge supposed to run.
pythonPlatform = "Linux"
